+++ KlimaKar/urls.py
@@ -3,7 +3,7 @@ from django.contrib import admin
 
 from ariadne.contrib.django.views import GraphQLView
 
-from KlimaKar.views import HomeView, SendIssueView, SendSMSView, ChangeLogView
+from KlimaKar.views import HomeView, SendIssueView, ChangeLogView
 from KlimaKar.graphql import schema
 
 urlpatterns = [
@@ -20,7 +20,6 @@ urlpatterns = [
     path("szukaj/", include("apps.search.urls")),
     path("tiles/", include("tiles.urls")),
     path("send_issue", SendIssueView.as_view(), name="send_issue"),
-    path("send_sms", SendSMSView.as_view(), name="send_sms"),
     path("lista-zmian", ChangeLogView.as_view(), name="changelog"),
     path("django-rq/", include("django_rq.urls")),
     path(