+++ KlimaKar/views.py
@@ -1,5 +1,3 @@
-import unicodedata
-
 from django.views.generic import TemplateView, View
 from django.http import JsonResponse, HttpResponseBadRequest
 from django.urls import reverse
@@ -8,11 +6,10 @@ from django.core.exceptions import ImproperlyConfigured
 from django_tables2 import SingleTableView
 from dal import autocomplete
 from github import Github
-from smsapi.client import SmsApiPlClient
 
 from KlimaKar import settings
 from KlimaKar.forms import IssueForm
-from KlimaKar.email import get_email_message, mail_admins
+from KlimaKar.email import get_email_message
 
 
 class HomeView(TemplateView):
@@ -359,7 +356,7 @@ class CustomSelect2QuerySetView(autocomplete.Select2QuerySetView):
 
         result = self.create_object(text)
 
-        return JsonResponse({"id": result.pk, "text": str(result),})
+        return JsonResponse({"id": result.pk, "text": str(result)})
 
     def get_results(self, context):
         return [
@@ -448,38 +445,6 @@ class FilteredSingleTableView(SingleTableView):
             )
         else:
             return super().get(request, args, kwargs)
-
-
-class SendSMSView(View):
-    def post(self, request, *args, **kwargs):
-        phone = request.POST.get("phone")
-        message = request.POST.get("message")
-        message = self._strip_accents(message)
-
-        if not phone or not message or not len(phone) == 9:
-            return JsonResponse(
-                {"status": "error", "message": "Coś poszło nie tak. Spróbuj ponownie."},
-                status=400,
-            )
-
-        client = SmsApiPlClient(access_token=settings.SMSAPI_TOKEN)
-        if int(client.account.balance().pro_count) < settings.SMSAPI_LOW_BALANCE_COUNT:
-            mail_admins("SMSAPI low balance", str(client.account.balance()))
-        client.sms.send(to=phone, message=message)
-        return JsonResponse(
-            {"status": "success", "message": "Wiadomość została wysłana.",}, status=200
-        )
-
-    def _strip_accents(self, text):
-        return (
-            "".join(
-                c
-                for c in unicodedata.normalize("NFKD", text)
-                if unicodedata.category(c) != "Mn"
-            )
-            .replace("ł", "l")
-            .replace("Ł", "L")
-        )
 
 
 class ChangeLogView(TemplateView):