+++ apps/commission/urls.py
@@ -150,4 +150,9 @@ urlpatterns = [
         views.CommissionAutocomplete.as_view(),
         name="commission_autocomplete",
     ),
+    path(
+        "sms_notification",
+        views.SendSMSNotificationView.as_view(),
+        name="send_sms_notification",
+    ),
 ]